+++ AmigaOS_NDK_3.1/Includes_Libs/include_h/clib/expansion_protos.h
@@ -17,32 +17,32 @@
 /*--- functions in V33 or higher (Release 1.2) ---*/
 void AddConfigDev( struct ConfigDev *configDev );
 /*--- functions in V36 or higher (Release 2.0) ---*/
-BOOL AddBootNode( long bootPri, unsigned long flags,
+BOOL AddBootNode( int32_t bootPri, uint32_t flags,
 	struct DeviceNode *deviceNode, struct ConfigDev *configDev );
 /*--- functions in V33 or higher (Release 1.2) ---*/
-void AllocBoardMem( unsigned long slotSpec );
+void AllocBoardMem( uint32_t slotSpec );
 struct ConfigDev *AllocConfigDev( void );
-APTR AllocExpansionMem( unsigned long numSlots, unsigned long slotAlign );
+APTR AllocExpansionMem( uint32_t numSlots, uint32_t slotAlign );
 void ConfigBoard( APTR board, struct ConfigDev *configDev );
 void ConfigChain( APTR baseAddr );
 struct ConfigDev *FindConfigDev( struct ConfigDev *oldConfigDev,
-	long manufacturer, long product );
-void FreeBoardMem( unsigned long startSlot, unsigned long slotSpec );
+	int32_t manufacturer, int32_t product );
+void FreeBoardMem( uint32_t startSlot, uint32_t slotSpec );
 void FreeConfigDev( struct ConfigDev *configDev );
-void FreeExpansionMem( unsigned long startSlot, unsigned long numSlots );
-UBYTE ReadExpansionByte( APTR board, unsigned long offset );
+void FreeExpansionMem( uint32_t startSlot, uint32_t numSlots );
+UBYTE ReadExpansionByte( APTR board, uint32_t offset );
 void ReadExpansionRom( APTR board, struct ConfigDev *configDev );
 void RemConfigDev( struct ConfigDev *configDev );
-void WriteExpansionByte( APTR board, unsigned long offset,
-	unsigned long byte );
+void WriteExpansionByte( APTR board, uint32_t offset,
+	uint32_t byte );
 void ObtainConfigBinding( void );
 void ReleaseConfigBinding( void );
 void SetCurrentBinding( struct CurrentBinding *currentBinding,
-	unsigned long bindingSize );
+	uint32_t bindingSize );
 ULONG GetCurrentBinding( struct CurrentBinding *currentBinding,
-	unsigned long bindingSize );
+	uint32_t bindingSize );
 struct DeviceNode *MakeDosNode( APTR parmPacket );
-BOOL AddDosNode( long bootPri, unsigned long flags,
+BOOL AddDosNode( int32_t bootPri, uint32_t flags,
 	struct DeviceNode *deviceNode );
 /*--- functions in V36 or higher (Release 2.0) ---*/
 #endif	 /* CLIB_EXPANSION_PROTOS_H */
